added default Room.
changed the main page view after login depending on the user's default room.
