+++ frontend/app/(app)/room/[id].tsx
@@ -14,6 +14,8 @@ import { useLocalSearchParams, router } from 'expo-router';
 import { useRoom } from '../../../src/store/roomContext';
 import { useAuth } from '../../../src/store/authContext';
 import UserSearchInput from '../../../src/components/rooms/UserSearchInput';
+import * as Clipboard from 'expo-clipboard';
+
 
 type User = {
   _id: string;
@@ -22,14 +24,14 @@ type User = {
 
 export default function RoomDetailScreen() {
   const { id } = useLocalSearchParams<{ id: string }>();
-  const { user,token } = useAuth();
-  const { currentRoom, fetchRoom, isLoading, inviteUsers, error, clearError } = useRoom();
+  const { user, token, logout } = useAuth();
+  const { rooms, currentRoom, fetchRoom, isLoading, isLoading: roomsLoading, inviteUsers, error, clearError } = useRoom();
   const [selectedUsers, setSelectedUsers] = useState<User[]>([]);
   const [isInviting, setIsInviting] = useState(false);
   const [showInviteForm, setShowInviteForm] = useState(false);
   const [localLoading, setLocalLoading] = useState(true);
   const [loadingError, setLoadingError] = useState<string | null>(null);
-  
+
   useEffect(() => {
     const loadRoomData = async () => {
       if (!id) {
@@ -37,12 +39,12 @@ export default function RoomDetailScreen() {
         setLocalLoading(false);
         return;
       }
-      
+
       if (!token) {
         console.log('Token not ready yet. Waiting to fetch room...');
         return; // 👈 WAIT until token exists
       }
-      
+
       try {
         console.log('Fetching room with ID:', id);
         await fetchRoom(id);
@@ -54,25 +56,42 @@ export default function RoomDetailScreen() {
         setLocalLoading(false);
       }
     };
-    
+
     loadRoomData();
   }, [id, token]); // 👈 depend on BOTH id and token
-  
-  
+
+  const handleCreateRoom = () => {
+    router.push('/(app)/create-room');
+  };
+
+  const handleJoinRoom = () => {
+    router.push('/(app)/join-room');
+  };
+
+  const handleLogout = async () => {
+    try {
+      await logout();
+      Alert.alert('Success', 'You have been logged out');
+    } catch (error) {
+      console.error('Logout error:', error);
+      Alert.alert('Error', 'Something went wrong during logout');
+    }
+  };
+
   const getMemberUsername = (member: any) => {
     if (!member) return 'Unknown';
-    
+
     if (typeof member === 'string') {
       return 'User';  // We only have the ID
     }
-    
+
     if (member._id) {
       return member.username || 'Unnamed';
     }
-    
+
     return 'Unknown';
   };
-  
+
   const handleSelectUser = (user: User) => {
     setSelectedUsers(prev => {
       if (prev.some(u => u._id === user._id)) {
@@ -81,36 +100,34 @@ export default function RoomDetailScreen() {
       return [...prev, user];
     });
   };
-  
+
   const handleRemoveUser = (userId: string) => {
     setSelectedUsers(prev => prev.filter(user => user._id !== userId));
   };
-  
+
   const handleInviteUsers = async () => {
     if (!currentRoom || isInviting || selectedUsers.length === 0) return;
-    
+
     setIsInviting(true);
     clearError();
-    
+
     try {
       const result = await inviteUsers(
         currentRoom._id,
         selectedUsers.map(user => user.username)
       );
-      
+
       if (result) {
         const { invited, invalid } = result;
-        
-        let message = `${invited.length} ${
-          invited.length === 1 ? 'user' : 'users'
-        } invited successfully.`;
-        
+
+        let message = `${invited.length} ${invited.length === 1 ? 'user' : 'users'
+          } invited successfully.`;
+
         if (invalid.length > 0) {
-          message += ` ${invalid.length} ${
-            invalid.length === 1 ? 'invitation' : 'invitations'
-          } could not be sent.`;
+          message += ` ${invalid.length} ${invalid.length === 1 ? 'invitation' : 'invitations'
+            } could not be sent.`;
         }
-        
+
         Alert.alert('Success', message);
         setSelectedUsers([]);
         setShowInviteForm(false);
@@ -122,19 +139,23 @@ export default function RoomDetailScreen() {
       setIsInviting(false);
     }
   };
-  
+
+  const [copied, setCopied] = useState(false);
+
   const handleShareRoomCode = async () => {
     if (!currentRoom) return;
-    
+  
     try {
-      await Share.share({
-        message: `Join my room in Katchup! Room code: ${currentRoom.code}`,
-      });
+      await Clipboard.setStringAsync(currentRoom.code);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 1000); 
     } catch (error) {
-      console.error('Share error:', error);
+      console.error('Clipboard copy error:', error);
     }
   };
   
+  
+
   // Show loading state while fetching data
   if (localLoading || isLoading) {
     return (
@@ -144,7 +165,7 @@ export default function RoomDetailScreen() {
       </SafeAreaView>
     );
   }
-  
+
   // Show error state
   if (loadingError || error || !currentRoom) {
     return (
@@ -161,36 +182,41 @@ export default function RoomDetailScreen() {
       </SafeAreaView>
     );
   }
-  
+
   return (
+
     <SafeAreaView className="flex-1 bg-gray-50">
       <ScrollView className="flex-1 p-4">
         <Text className="text-2xl font-bold text-center mb-2">{currentRoom.name}</Text>
-        
         <View className="bg-white rounded-lg shadow-sm p-4 mb-6">
           <View className="flex-row justify-between items-center mb-4">
             <Text className="text-gray-600">Room Code</Text>
             <Text className="font-bold">{currentRoom.code}</Text>
           </View>
-          
+
           <Pressable
             className="bg-blue-500 p-3 rounded-lg"
             onPress={handleShareRoomCode}
           >
             <Text className="text-white text-center font-medium">Share Room Code</Text>
           </Pressable>
+          {copied && (
+            <Text className="text-green-500 text-center mt-2">
+              Room code copied to clipboard!
+            </Text>
+          )}
         </View>
-        
+
         {/* Members List */}
         <View className="bg-white rounded-lg shadow-sm p-4 mb-6">
           <Text className="text-gray-800 font-medium text-lg mb-3">
             Members ({currentRoom.members?.length || 0})
           </Text>
-          
+
           {currentRoom.members && currentRoom.members.length > 0 ? (
             currentRoom.members.map((member, index) => (
-              <View 
-                key={`member-${index}`} 
+              <View
+                key={`member-${index}`}
                 className="flex-row justify-between py-2 border-b border-gray-100"
               >
                 <Text className="text-gray-800">{getMemberUsername(member)}</Text>
@@ -200,24 +226,24 @@ export default function RoomDetailScreen() {
             <Text className="text-gray-500 italic">No members found</Text>
           )}
         </View>
-        
+
         {/* Pending Invitations */}
         {currentRoom.invitations && currentRoom.invitations.filter(inv => inv.status === 'pending').length > 0 && (
           <View className="bg-white rounded-lg shadow-sm p-4 mb-6">
             <Text className="text-gray-800 font-medium text-lg mb-3">
               Pending Invitations (
-                {currentRoom.invitations.filter(inv => inv.status === 'pending').length}
+              {currentRoom.invitations.filter(inv => inv.status === 'pending').length}
               )
             </Text>
-            
+
             {currentRoom.invitations.filter(inv => inv.status === 'pending').map((invitation, index) => {
-              const inviteeName = typeof invitation.user === 'string' 
-                ? 'User' 
+              const inviteeName = typeof invitation.user === 'string'
+                ? 'User'
                 : invitation.user.username || 'Unnamed';
-              
+
               return (
-                <View 
-                  key={`invitation-${index}`} 
+                <View
+                  key={`invitation-${index}`}
                   className="flex-row justify-between py-2 border-b border-gray-100"
                 >
                   <Text className="text-gray-800">{inviteeName}</Text>
@@ -227,7 +253,7 @@ export default function RoomDetailScreen() {
             })}
           </View>
         )}
-        
+
         {/* Invite Users Form */}
         {!showInviteForm ? (
           <Pressable
@@ -243,17 +269,17 @@ export default function RoomDetailScreen() {
             <Text className="text-gray-800 font-medium text-lg mb-3">
               Invite Users
             </Text>
-            
+
             <UserSearchInput
               selectedUsers={selectedUsers}
               onSelectUser={handleSelectUser}
               onRemoveUser={handleRemoveUser}
             />
-            
+
             {error && (
               <Text className="text-red-500 mb-4">{error}</Text>
             )}
-            
+
             <View className="flex-row justify-end space-x-3 mt-4">
               <Pressable
                 className="bg-gray-200 px-4 py-2 rounded-lg"
@@ -265,11 +291,10 @@ export default function RoomDetailScreen() {
               >
                 <Text className="text-gray-700">Cancel</Text>
               </Pressable>
-              
+
               <Pressable
-                className={`bg-red-500 px-4 py-2 rounded-lg ${
-                  isInviting || selectedUsers.length === 0 ? 'opacity-50' : ''
-                }`}
+                className={`bg-red-500 px-4 py-2 rounded-lg ${isInviting || selectedUsers.length === 0 ? 'opacity-50' : ''
+                  }`}
                 onPress={handleInviteUsers}
                 disabled={isInviting || selectedUsers.length === 0}
               >
@@ -282,7 +307,30 @@ export default function RoomDetailScreen() {
             </View>
           </View>
         )}
-        
+        {/* Rooms List or Login/Create/Join Options */}
+        <View className="flex-row space-x-2 mt-4">
+          <Pressable
+            className="bg-red-500 p-3 rounded-lg flex-1"
+            onPress={handleCreateRoom}
+          >
+            <Text className="text-white text-center font-medium">
+              Create Room
+            </Text>
+          </Pressable>
+          <Pressable
+            className="bg-blue-500 p-3 rounded-lg flex-1"
+            onPress={handleJoinRoom}
+          >
+            <Text className="text-white text-center font-medium">
+              Join Room
+            </Text>
+          </Pressable>
+          <Pressable
+            onPress={handleLogout}
+            className="bg-gray-500 p-3 rounded-lg flex-1">
+            <Text className='text-white text-center font-medium'>Logout</Text>
+          </Pressable>
+        </View>
         {/* Back Button */}
         <Pressable
           className="p-4 mt-2"
