+++ frontend/package.json
@@ -21,6 +21,7 @@
     "@react-navigation/native": "^7.0.14",
     "expo": "~52.0.46",
     "expo-blur": "~14.0.3",
+    "expo-clipboard": "^7.0.1",
     "expo-constants": "~17.0.8",
     "expo-font": "~13.0.4",
     "expo-haptics": "~14.0.1",
